+++ packages/api/src/client/lexicons.ts
@@ -891,7 +891,7 @@ export const schemaDict = {
       labelValueDefinition: {
         type: 'object',
         description:
-          'Declares a label value and its expected interpertations and behaviors.',
+          'Declares a label value and its expected interpretations and behaviors.',
         required: ['identifier', 'severity', 'blurs', 'locales'],
         properties: {
           identifier: {
@@ -4359,6 +4359,9 @@ export const schemaDict = {
         description: 'A word that the account owner has muted.',
         required: ['value', 'targets'],
         properties: {
+          id: {
+            type: 'string',
+          },
           value: {
             type: 'string',
             description: 'The muted word itself.',
@@ -4372,6 +4375,19 @@ export const schemaDict = {
               type: 'ref',
               ref: 'lex:app.bsky.actor.defs#mutedWordTarget',
             },
+          },
+          actorTarget: {
+            type: 'string',
+            description:
+              'Groups of users to apply the muted word to. If undefined, applies to all users.',
+            knownValues: ['all', 'exclude-following'],
+            default: 'all',
+          },
+          expiresAt: {
+            type: 'string',
+            format: 'datetime',
+            description:
+              'The date and time at which the muted word will expire and no longer be applied.',
           },
         },
       },
@@ -5484,7 +5500,7 @@ export const schemaDict = {
           feedContext: {
             type: 'string',
             description:
-              'Context on a feed item that was orginally supplied by the feed generator on getFeedSkeleton.',
+              'Context on a feed item that was originally supplied by the feed generator on getFeedSkeleton.',
             maxLength: 2000,
           },
         },